producer/requirements.txt
azure-eventgrid==4.1.0
azure-core==1.17.0